requirements.txt
Flask==2.3.3
flask-cors==5.0.0
Pillow==10.4.0
PyMuPDF==1.24.10
mysql-connector-python==9.0.0
pytesseract==0.3.13
